Fixed issue with styled components.
Constantly updating background and font colors was causing enormous numbers of styles to be generated and added to the head tag in index.html. This rapidly bogged down the performance of the app. Now most dynamic style changes are handled in the style key of the attrs object (which is equivalent of setting the style inline  on the element) or it is set inline on the element. Consequently, no additional styles are being added in the head tag and performance doesn't degrade. Hurrah!